implement generic struct and slice comparison for testing
TestReq now may feature expectedData which can be a struct or a slice of struct (which is all we need right now but should be easily expandable). This is automatically compared to the data returned by the request. If any struct field is missing in expected data, that's not a problem because fields that contain their types zero value are skipped (empty strings, 0, nil etc.).