refactor structure for data access
I think "what data part(s) do we want to add/get/modify/delete" is a question for handlers, not the db client. This way we keep the db more versatile and this behavior would't change for other dbs (=no reimplementation needed). Also for some things we will need authorization information, which we already have in our handlers (context locals) but would have to pass down for handling in db package